Beta Release 60 - Includes up to Developer Release 7493

+++ scripts/system/edit.js
@@ -419,7 +419,7 @@ var toolBar = (function () {
         var createButtonIconRsrc = (hasRezPermissions ? CREATE_ENABLED_ICON : CREATE_DISABLED_ICON);
         tablet = Tablet.getTablet("com.highfidelity.interface.tablet.system");
         activeButton = tablet.addButton({
-            captionColorOverride: hasRezPermissions ? "" : "#888888",
+            captionColor: hasRezPermissions ? "#ffffff" : "#888888",
             icon: createButtonIconRsrc,
             activeIcon: "icons/tablet-icons/edit-a.svg",
             text: "CREATE",
@@ -792,7 +792,7 @@ function handleDomainChange() {
     var hasRezPermissions = (Entities.canRez() || Entities.canRezTmp() || Entities.canRezCertified() || Entities.canRezTmpCertified());
     createButton.editProperties({
         icon: (hasRezPermissions ? CREATE_ENABLED_ICON : CREATE_DISABLED_ICON),
-        captionColorOverride: (hasRezPermissions ? "" : "#888888"),
+        captionColor: (hasRezPermissions ? "#ffffff" : "#888888"),
     });
 }
 
@@ -1424,24 +1424,29 @@ function deleteSelectedEntities() {
         for (var i = 0; i < newSortedSelection.length; i++) {
             var entityID = newSortedSelection[i];
             var initialProperties = SelectionManager.savedProperties[entityID];
-            var children = Entities.getChildrenIDs(entityID);
-            var childList = [];
-            recursiveDelete(children, childList, deletedIDs);
-            savedProperties.push({
-                entityID: entityID,
-                properties: initialProperties,
-                children: childList
-            });
-            deletedIDs.push(entityID);
-            Entities.deleteEntity(entityID);
+            if (!initialProperties.locked) {
+                var children = Entities.getChildrenIDs(entityID);
+                var childList = [];
+                recursiveDelete(children, childList, deletedIDs);
+                savedProperties.push({
+                    entityID: entityID,
+                    properties: initialProperties,
+                    children: childList
+                });
+                deletedIDs.push(entityID);
+                Entities.deleteEntity(entityID);
+            }
         }
-        SelectionManager.clearSelections();
-        pushCommandForSelections([], savedProperties);
 
-        entityListTool.webView.emitScriptEvent(JSON.stringify({
-            type: "deleted",
-            ids: deletedIDs
-        }));
+        if (savedProperties.length > 0) {
+            SelectionManager.clearSelections();
+            pushCommandForSelections([], savedProperties);
+
+            entityListTool.webView.emitScriptEvent(JSON.stringify({
+                type: "deleted",
+                ids: deletedIDs
+            }));
+        }
     }
 }
